start of geometry glueing tool (#10)
squashing again because it's a bit of a mess

* start of geometry glueing tool
* add a brep_flatten tool
* use git lfs to track larger test-data files
* add merge help to README
* describe brep file format
* example of creating input for gluer regression test
* start automatic testing of shape merge code
* more compiler warnings and turn them into errors when building in CI
* give an option for generating code coverage

currently hardcoded to Clang style "source-based code coverage"

* lots of other small/misc changes

+++ src/geometry.hpp
@@ -1,3 +1,4 @@
+#include <sys/types.h>
 #include <vector>
 #include <ostream>
 
@@ -25,7 +26,7 @@ struct document {
 	size_t count_invalid_shapes() const;
 
 	// only integer indexes supported at the moment, returns -1 if invalid
-	int lookup_solid(const std::string &str) const;
+	ssize_t lookup_solid(const std::string &str) const;
 };
 
 class BOPAlgo_PaveFiller;
